Whatever packages/files are related to data is moved to new data package
